+++ src/packages/ssc.rs
@@ -541,6 +541,8 @@ mod tests {
                 distribution_date: None,
                 files: vec![],
                 description_lines: vec![],
+                requires: vec![],
+                stata_version: None,
             },
             files: vec![
                 DownloadedFile {